@@ -1,8 +1,8 @@
 plugins {
-	id 'fabric-loom' version '1.1-SNAPSHOT'
+	id 'fabric-loom' version '1.9-SNAPSHOT'
 }
 
-sourceCompatibility = targetCompatibility = JavaVersion.VERSION_17
+sourceCompatibility = targetCompatibility = JavaVersion.VERSION_21
 
 archivesBaseName = project.archives_base_name
 version = project.mod_version
@@ -16,6 +16,11 @@ repositories {
 	maven { url "https://maven.seedfinding.com/" }
 	maven { url "https://maven-snapshots.seedfinding.com/" }
 	maven { url 'https://jitpack.io' }
+	maven { url 'https://maven.duti.dev/releases' }
+}
+
+loom {
+	accessWidenerPath = file("src/main/resources/meteor-rejects.accesswidener")
 }
 
 configurations {
@@ -24,13 +29,18 @@ configurations {
 }
 
 dependencies {
+	// This will make it work on most platforms. It automatically chooses the right dependencies at runtime.
+	extraLibs('dev.duti.acheong:cubiomes:1.22.3') { transitive = false }
+	extraLibs('dev.duti.acheong:cubiomes:1.22.3:linux64') { transitive = false }
+	extraLibs('dev.duti.acheong:cubiomes:1.22.3:osx') { transitive = false }
+	extraLibs('dev.duti.acheong:cubiomes:1.22.3:windows64') { transitive = false }
 	// To change the versions see the gradle.properties file
 	minecraft "com.mojang:minecraft:${project.minecraft_version}"
 	mappings "net.fabricmc:yarn:${project.yarn_version}:v2"
 	modImplementation "net.fabricmc:fabric-loader:${project.loader_version}"
 
-	modImplementation("meteordevelopment:meteor-client:${project.meteor_version}-SNAPSHOT")
-	modImplementation "baritone:fabric:${project.minecraft_version}-SNAPSHOT"
+	modImplementation("meteordevelopment:meteor-client:${project.minecraft_version}-SNAPSHOT")
+    modCompileOnly "meteordevelopment:baritone:${project.baritone_version}-SNAPSHOT"
 
 	// seed .locate and ore sim
 	extraLibs('com.seedfinding:mc_math:ffd2edcfcc0d18147549c88cc7d8ec6cf21b5b91') { transitive = false }
@@ -40,16 +50,12 @@ dependencies {
 	extraLibs('com.seedfinding:mc_biome:41a42cb9019a552598f12089059538853e18ec78') { transitive = false }
 	extraLibs('com.seedfinding:mc_terrain:b4246cbd5880c4f8745ccb90e1b102bde3448126') { transitive = false }
 	extraLibs('com.seedfinding:mc_feature:919b7e513cc1e87e029a9cd703fc4e2dc8686229') { transitive = false }
-	
+
 	// seedcracker api
 	implementation (include('com.github.19MisterX98.SeedcrackerX:seedcrackerx-api:2.10.1')) {transitive = false}
 //	implementation (include('com.github.19MisterX98.SeedcrackerX:seedcrackerx-api:master-SNAPSHOT')) {transitive = false}
 
 	configurations.implementation.extendsFrom(configurations.extraLibs)
-}
-
-loom {
-    accessWidenerPath = file("src/main/resources/meteor-rejects.accesswidener")
 }
 
 processResources {
@@ -65,7 +71,7 @@ processResources {
 
 jar {
 	from("LICENSE") {
-		rename { "${it}_${project.archivesBaseName}"}
+		rename { "${it}_${project.archivesBaseName}" }
 	}
 	from {
 		configurations.extraLibs.collect { it.isDirectory() ? it : zipTree(it) }
@@ -85,7 +91,7 @@ tasks.withType(JavaCompile).configureEach {
 	// The Minecraft launcher currently installs Java 8 for users, so your mod probably wants to target Java 8 too
 	// JDK 9 introduced a new way of specifying this that will make sure no newer classes or methods are used.
 	// We'll use that if it's available, but otherwise we'll use the older option.
-	def targetVersion = 17
+	def targetVersion = 21
 	if (JavaVersion.current().isJava9Compatible()) {
 		 it.options.release = targetVersion
 	}
