@@ -7,10 +7,14 @@ import meteordevelopment.meteorclient.settings.ItemListSetting;
 import meteordevelopment.meteorclient.settings.Setting;
 import meteordevelopment.meteorclient.settings.SettingGroup;
 import meteordevelopment.meteorclient.systems.modules.Module;
+import meteordevelopment.meteorclient.utils.Utils;
 import meteordevelopment.orbit.EventHandler;
 import net.minecraft.client.gui.screen.recipebook.RecipeResultCollection;
 import net.minecraft.item.Item;
-import net.minecraft.recipe.Recipe;
+import net.minecraft.item.ItemStack;
+import net.minecraft.recipe.RecipeDisplayEntry;
+import net.minecraft.recipe.display.RecipeDisplay;
+import net.minecraft.recipe.display.SlotDisplayContexts;
 import net.minecraft.screen.CraftingScreenHandler;
 import net.minecraft.screen.slot.SlotActionType;
 
@@ -24,7 +28,7 @@ public class AutoCraft extends Module {
     private final Setting<List<Item>> items = sgGeneral.add(new ItemListSetting.Builder()
         .name("items")
         .description("Items you want to get crafted.")
-        .defaultValue(Arrays.asList())
+        .defaultValue(List.of())
         .build()
     );
 
@@ -55,11 +59,11 @@ public class AutoCraft extends Module {
     
     @EventHandler
     private void onTick(TickEvent.Post event) {
-        if (mc.interactionManager == null) return;
+        if (!Utils.canUpdate() || mc.interactionManager == null) return;
+
         if (items.get().isEmpty()) return;
 
         if (!(mc.player.currentScreenHandler instanceof CraftingScreenHandler)) return;
-        
 
         if (antiDesync.get()) 
             mc.player.getInventory().updateItems();
@@ -70,11 +74,19 @@ public class AutoCraft extends Module {
         List<Item> itemList = items.get();
         List<RecipeResultCollection> recipeResultCollectionList  = mc.player.getRecipeBook().getOrderedResults();
         for (RecipeResultCollection recipeResultCollection : recipeResultCollectionList) {
-            for (Recipe<?> recipe : recipeResultCollection.getRecipes(true)) {
-                if (!itemList.contains(recipe.getOutput(mc.world.getRegistryManager()).getItem())) continue;
-                mc.interactionManager.clickRecipe(currentScreenHandler.syncId, recipe, craftAll.get());
-                mc.interactionManager.clickSlot(currentScreenHandler.syncId, 0, 1,
-                        drop.get() ? SlotActionType.THROW : SlotActionType.QUICK_MOVE, mc.player);
+            // Get craftable recipes only
+            List<RecipeDisplayEntry> craftRecipes = recipeResultCollection.filter(RecipeResultCollection.RecipeFilterMode.CRAFTABLE);
+            for (RecipeDisplayEntry recipe : craftRecipes) {
+                RecipeDisplay recipeDisplay = recipe.display();
+                List<ItemStack> resultStacks = recipeDisplay.result().getStacks(SlotDisplayContexts.createParameters(mc.world));
+                for (ItemStack resultStack : resultStacks) {
+                    // Check if the result item is in the item list
+                    if (!itemList.contains(resultStack.getItem())) continue;
+
+                    mc.interactionManager.clickRecipe(currentScreenHandler.syncId, recipe.id(), craftAll.get());
+                    mc.interactionManager.clickSlot(currentScreenHandler.syncId, 0, 1,
+                            drop.get() ? SlotActionType.THROW : SlotActionType.QUICK_MOVE, mc.player);
+                }
             }
         }
     }
