@@ -6,9 +6,6 @@ import net.minecraft.command.CommandSource;
 import net.minecraft.network.packet.c2s.play.PlayerMoveC2SPacket;
 import net.minecraft.util.math.MathHelper;
 
-import static com.mojang.brigadier.Command.SINGLE_SUCCESS;
-import static meteordevelopment.meteorclient.MeteorClient.mc;
-
 public class CenterCommand extends Command {
     public CenterCommand() {
         super("center", "Centers the player on a block.");
@@ -20,7 +17,7 @@ public class CenterCommand extends Command {
             double x = MathHelper.floor(mc.player.getX()) + 0.5;
             double z = MathHelper.floor(mc.player.getZ()) + 0.5;
             mc.player.setPosition(x, mc.player.getY(), z);
-            mc.player.networkHandler.sendPacket(new PlayerMoveC2SPacket.PositionAndOnGround(mc.player.getX(), mc.player.getY(), mc.player.getZ(), mc.player.isOnGround()));
+            mc.player.networkHandler.sendPacket(new PlayerMoveC2SPacket.PositionAndOnGround(mc.player.getX(), mc.player.getY(), mc.player.getZ(), mc.player.isOnGround(), mc.player.horizontalCollision));
 
             return SINGLE_SUCCESS;
         }));
@@ -29,7 +26,7 @@ public class CenterCommand extends Command {
             double x = MathHelper.floor(mc.player.getX());
             double z = MathHelper.floor(mc.player.getZ());
             mc.player.setPosition(x, mc.player.getY(), z);
-            mc.player.networkHandler.sendPacket(new PlayerMoveC2SPacket.PositionAndOnGround(mc.player.getX(), mc.player.getY(), mc.player.getZ(), mc.player.isOnGround()));
+            mc.player.networkHandler.sendPacket(new PlayerMoveC2SPacket.PositionAndOnGround(mc.player.getX(), mc.player.getY(), mc.player.getZ(), mc.player.isOnGround(), mc.player.horizontalCollision));
 
             return SINGLE_SUCCESS;
         }));
