@@ -15,15 +15,18 @@ import meteordevelopment.starscript.compiler.Parser;
 import meteordevelopment.starscript.utils.Error;
 import meteordevelopment.starscript.utils.StarscriptError;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.gl.ShaderProgramKeys;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.screen.ChatScreen;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.screen.ingame.InventoryScreen;
 import net.minecraft.client.option.KeyBinding;
-import net.minecraft.client.render.GameRenderer;
+import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.util.InputUtil;
 import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.component.DataComponentTypes;
 import net.minecraft.entity.Entity;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.Saddleable;
 import net.minecraft.entity.mob.EndermanEntity;
 import net.minecraft.entity.passive.AbstractHorseEntity;
@@ -36,6 +39,7 @@ import net.minecraft.network.packet.c2s.play.PlayerInteractEntityC2SPacket;
 import net.minecraft.text.Text;
 import net.minecraft.util.Hand;
 import net.minecraft.util.Identifier;
+import net.minecraft.util.PlayerInput;
 import net.minecraft.util.math.MathHelper;
 import org.joml.Vector2f;
 import org.lwjgl.glfw.GLFW;
@@ -60,7 +64,7 @@ public class InteractionScreen extends Screen {
     private final Map<String, Consumer<Entity>> functions;
     private final Map<String, String> msgs;
 
-    private final Identifier GUI_ICONS_TEXTURE = new Identifier("textures/gui/icons.png");
+    private final Identifier GUI_ICONS_TEXTURE = Identifier.of("textures/gui/icons.png");
 
     private final StaticListener shiftListener = new StaticListener();
 
@@ -91,30 +95,29 @@ public class InteractionScreen extends Screen {
             closeScreen();
             client.setScreen(new StatsScreen(e));
         });
-        if (entity instanceof PlayerEntity) {
-            functions.put("Open Inventory", (Entity e) -> {
+        switch (entity) {
+            case PlayerEntity playerEntity -> functions.put("Open Inventory", (Entity e) -> {
                 closeScreen();
                 client.setScreen(new InventoryScreen((PlayerEntity) e));
             });
-        } else if (entity instanceof AbstractHorseEntity) {
-            functions.put("Open Inventory", (Entity e) -> {
+            case AbstractHorseEntity abstractHorseEntity -> functions.put("Open Inventory", (Entity e) -> {
                 closeScreen();
                 if (client.player.isRiding()) {
-                    client.player.networkHandler.sendPacket(new PlayerInputC2SPacket(0, 0, false, true));
+//                    client.player.networkHandler.sendPacket(new PlayerInputC2SPacket(0, 0, false, true));
+                    client.player.networkHandler.sendPacket(new PlayerInputC2SPacket(new PlayerInput(false, false, false, false, false, true, false)));
+
                 }
                 client.player.networkHandler.sendPacket(PlayerInteractEntityC2SPacket.interact(entity, true, Hand.MAIN_HAND));
                 client.player.setSneaking(false);
             });
-        } else if (entity instanceof StorageMinecartEntity) {
-            functions.put("Open Inventory", (Entity e) -> {
+            case StorageMinecartEntity storageMinecartEntity -> functions.put("Open Inventory", (Entity e) -> {
                 closeScreen();
                 client.player.networkHandler.sendPacket(PlayerInteractEntityC2SPacket.interact(entity, true, Hand.MAIN_HAND));
             });
-        } else {
-            functions.put("Open Inventory", (Entity e) -> {
+            case null, default -> functions.put("Open Inventory", (Entity e) -> {
                 closeScreen();
                 ItemStack container = new ItemStack(Items.CHEST);
-                container.setCustomName(e.getName());
+                container.set(DataComponentTypes.CUSTOM_NAME, e.getName());
                 client.setScreen(new PeekScreen(container, getInventory(e)));
             });
         }
@@ -184,24 +187,27 @@ public class InteractionScreen extends Screen {
             } catch (NullPointerException ex) {
             }
         }
-        if (Saddleable.class.isInstance(e)) {
+        if (e instanceof Saddleable) {
             if (((Saddleable) e).isSaddled()) {
                 stack[index[0]] = Items.SADDLE.getDefaultStack();
                 index[0]++;
             }
         }
-        e.getHandItems().forEach(itemStack -> {
+        LivingEntity a = (LivingEntity) e;
+        a.getHandItems().forEach(itemStack -> {
             if (itemStack != null) {
                 stack[index[0]] = itemStack;
                 index[0]++;
             }
         });
-        e.getArmorItems().forEach(itemStack -> {
+
+        a.getArmorItems().forEach(itemStack -> {
             if (itemStack != null) {
                 stack[index[0]] = itemStack;
                 index[0]++;
             }
         });
+
         for (int i = index[0]; i < 27; i++) stack[i] = Items.AIR.getDefaultStack();
         return stack;
     }
@@ -246,12 +252,12 @@ public class InteractionScreen extends Screen {
         MatrixStack matrix = context.getMatrices();
         // Fake crosshair stuff
         RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.setShader(GameRenderer::getPositionTexProgram);
+        RenderSystem.setShader(ShaderProgramKeys.POSITION_TEX);
         RenderSystem.enableBlend();
         RenderSystem.blendFuncSeparate(GlStateManager.SrcFactor.ONE_MINUS_DST_COLOR,
                 GlStateManager.DstFactor.ONE_MINUS_SRC_COLOR, GlStateManager.SrcFactor.ONE,
                 GlStateManager.DstFactor.ZERO);
-        context.drawTexture(GUI_ICONS_TEXTURE, crosshairX - 8, crosshairY - 8, 0, 0, 15, 15);
+        context.drawTexture(RenderLayer::getGuiTextured, GUI_ICONS_TEXTURE,  crosshairX - 8, crosshairY - 8, 0, 0, 15, 15, 256, 256);
 
         drawDots(context, (int) (Math.min(height, width) / 2 * 0.75), mouseX, mouseY);
         matrix.scale(2f, 2f, 1f);
@@ -285,7 +291,7 @@ public class InteractionScreen extends Screen {
 
     private void drawDots(DrawContext context, int radius, int mouseX, int mouseY) {
         ArrayList<Point> pointList = new ArrayList<Point>();
-        String cache[] = new String[functions.size()];
+        String[] cache = new String[functions.size()];
         double lowestDistance = Double.MAX_VALUE;
         int i = 0;
 
