@@ -108,7 +108,7 @@ public class LegacyServerFinderScreen extends WindowScreen {
                         updatePingers(pingers);
                     }
                 }
-            while (pingers.size() > 0) {
+            while (!pingers.isEmpty()) {
                 if (state == ServerFinderState.CANCELLED)
                     return;
 
@@ -156,7 +156,7 @@ public class LegacyServerFinderScreen extends WindowScreen {
                     if (!isServerInList(pingers.get(i).getServerIP())) {
                         multiplayerScreen.getServerList()
                                 .add(new ServerInfo("Server discovery " + working,
-                                        pingers.get(i).getServerIP(), false), false);
+                                        pingers.get(i).getServerIP(), ServerInfo.ServerType.OTHER), false);
                         multiplayerScreen.getServerList().saveFile();
                         ((MultiplayerScreenAccessor) multiplayerScreen).getServerListWidget()
                                 .setSelected(null);
