@@ -1,17 +1,18 @@
 package anticope.rejects.mixin;
 
-import net.minecraft.network.packet.c2s.play.PlayerMoveC2SPacket;
+import net.minecraft.network.packet.c2s.play.VehicleMoveC2SPacket;
+import net.minecraft.util.math.Vec3d;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.gen.Accessor;
+import org.spongepowered.asm.mixin.gen.Invoker;
 
-@Mixin(PlayerMoveC2SPacket.class)
+@Mixin(VehicleMoveC2SPacket.class)
 public interface VehicleMoveC2SPacketAccessor {
-    @Mutable
-    @Accessor("x")
-    void setX(double x);
+    @Accessor("position")
+    Vec3d getPosition();
 
-    @Mutable
-    @Accessor("z")
-    void setZ(double z);
+    @Invoker("<init>")
+    static VehicleMoveC2SPacket create(Vec3d position, float yaw, float pitch, boolean onGround) {
+        throw new AssertionError();
+    }
 }
