@@ -17,9 +17,9 @@ public class ModuleMixin {
     @Mutable @Shadow public String name;
 
     @Mutable @Shadow public String title;
-
-    @Inject(method = "<init>", at = @At("TAIL"))
-    private void onInit(Category category, String name, String description, CallbackInfo info) {
+  
+    @Inject(method = "<init>*", at = @At("TAIL"))
+    private void onInit(Category category, String name, String description, String[] aliases, CallbackInfo info) {
         if (RejectsConfig.get().duplicateModuleNames) {
             this.name = RejectsUtils.getModuleName(name);
             this.title = Utils.nameToTitle(this.name);
