@@ -75,7 +75,7 @@ public class AntiVanish extends Module {
     @EventHandler
     private void onPacket(PacketEvent.Receive event) {
         if (mode.get() == Mode.RealJoinMessage && event.packet instanceof CommandSuggestionsS2CPacket packet) {
-            if (completionIDs.contains(packet.getCompletionId())) {
+            if (completionIDs.contains(packet.id())) {
                 var lastUsernames = completionPlayerCache.stream().toList();
 
                 completionPlayerCache = packet.getSuggestions().getList().stream()
@@ -88,6 +88,8 @@ public class AntiVanish extends Module {
 
                 for (String playerName : completionPlayerCache) {
                     if (Objects.equals(playerName, mc.player.getName().getString())) continue;
+                    if (playerName.contains(" ")) continue;
+                    if (playerName.length() < 3 || playerName.length() > 16) continue;
                     if (joinedOrQuit.test(playerName)) {
                         info("Player joined: " + playerName);
                     }
@@ -95,12 +97,14 @@ public class AntiVanish extends Module {
 
                 for (String playerName : lastUsernames) {
                     if (Objects.equals(playerName, mc.player.getName().getString())) continue;
+                    if (playerName.contains(" ")) continue;
+                    if (playerName.length() < 3 || playerName.length() > 16) continue;
                     if (joinedOrQuit.test(playerName)) {
                         info("Player left: " + playerName);
                     }
                 }
 
-                completionIDs.remove(Integer.valueOf(packet.getCompletionId()));
+                completionIDs.remove(Integer.valueOf(packet.id()));
                 event.cancel();
             }
         }
@@ -124,6 +128,8 @@ public class AntiVanish extends Module {
                 for (UUID uuid : oldPlayers.keySet()) {
                     if (playerCache.containsKey(uuid)) continue;
                     String name = oldPlayers.get(uuid);
+                    if (name.contains(" ")) continue;
+                    if (name.length() < 3 || name.length() > 16) continue;
                     if (messageCache.stream().noneMatch(s -> s.contains(name))) {
                         warning(name + " has gone into vanish.");
                     }
