@@ -151,6 +151,8 @@ public class AutoFarm extends Module {
             else if (block == Blocks.POTATOES) item = Items.POTATO;
             else if (block == Blocks.BEETROOTS) item = Items.BEETROOT_SEEDS;
             else if (block == Blocks.NETHER_WART) item = Items.NETHER_WART;
+            else if (block == Blocks.PITCHER_CROP) item = Items.PITCHER_POD;
+            else if (block == Blocks.TORCHFLOWER) item = Items.TORCHFLOWER_SEEDS;
             if (item != null) replantMap.put(event.blockPos, item);
         }
     }
@@ -257,32 +259,40 @@ public class AutoFarm extends Module {
         if (block instanceof CropBlock cropBlock) {
             return cropBlock.isMature(state);
         } else if (block instanceof CocoaBlock cocoaBlock) {
-            return !cocoaBlock.hasRandomTicks(state);
+            return state.get(cocoaBlock.AGE) >= 2;
         } else if (block instanceof StemBlock) {
             return state.get(StemBlock.AGE) == StemBlock.MAX_AGE;
         } else if (block instanceof SweetBerryBushBlock sweetBerryBushBlock) {
-            return !sweetBerryBushBlock.hasRandomTicks(state);
+            return state.get(sweetBerryBushBlock.AGE) >= 2;
         } else if (block instanceof NetherWartBlock netherWartBlock) {
-            return !netherWartBlock.hasRandomTicks(state);
+            return state.get(netherWartBlock.AGE) >= 3;
+        } else if (block instanceof PitcherCropBlock pitcherCropBlock) {
+            return state.get(pitcherCropBlock.AGE) >= 4;
         }
         return true;
     }
 
     private boolean bonemealFilter(Block block) {
         return block instanceof CropBlock ||
-                block instanceof CocoaBlock ||
                 block instanceof StemBlock ||
                 block instanceof MushroomPlantBlock ||
-                block instanceof SweetBerryBushBlock ||
                 block instanceof AzaleaBlock ||
-                block instanceof SaplingBlock;
+                block instanceof SaplingBlock ||
+                block == Blocks.COCOA ||
+                block == Blocks.SWEET_BERRY_BUSH ||
+                block == Blocks.PITCHER_CROP ||
+                block == Blocks.TORCHFLOWER;
     }
 
     private boolean harvestFilter(Block block) {
         return block instanceof CropBlock ||
-                block instanceof GourdBlock ||
-                block instanceof NetherWartBlock ||
-                block instanceof SweetBerryBushBlock;
+                block == Blocks.PUMPKIN ||
+                block == Blocks.MELON ||
+                block == Blocks.NETHER_WART ||
+                block == Blocks.SWEET_BERRY_BUSH ||
+                block == Blocks.COCOA ||
+                block == Blocks.PITCHER_CROP ||
+                block == Blocks.TORCHFLOWER;
     }
 
     private boolean plantFilter(Item item) {
@@ -292,6 +302,8 @@ public class AutoFarm extends Module {
                 item == Items.BEETROOT_SEEDS ||
                 item == Items.PUMPKIN_SEEDS ||
                 item == Items.MELON_SEEDS ||
-                item == Items.NETHER_WART;
+                item == Items.NETHER_WART ||
+                item == Items.PITCHER_POD ||
+                item == Items.TORCHFLOWER_SEEDS;
     }
 }
