@@ -142,9 +142,10 @@ public class ChorusExploit extends Module {
         if (event.packet instanceof PlayerPositionLookS2CPacket posPacket && ateChorus) {
             event.setCancelled(true);
             if (positionMode.get() == PositionMode.PosLook) {
-                cposX = posPacket.getX();
-                cposY = posPacket.getY();
-                cposZ = posPacket.getZ();
+                Vec3d pos = posPacket.change().position();
+                cposX = pos.x;
+                cposY = pos.y;
+                cposZ = pos.z;
                 gotPosition = true;
             }
         }
