@@ -3,6 +3,8 @@ package anticope.rejects.utils;
 import anticope.rejects.utils.seeds.Seed;
 import anticope.rejects.utils.seeds.Seeds;
 import baritone.api.BaritoneAPI;
+import cubitect.Cubiomes;
+
 import com.seedfinding.mcbiome.source.BiomeSource;
 import com.seedfinding.mcfeature.misc.SlimeChunk;
 import com.seedfinding.mcfeature.structure.*;
@@ -17,6 +19,8 @@ import meteordevelopment.meteorclient.utils.player.PlayerUtils;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
+import net.minecraft.component.DataComponentTypes;
+import net.minecraft.component.type.MapDecorationsComponent;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.mob.*;
 import net.minecraft.entity.passive.IronGolemEntity;
@@ -24,7 +28,6 @@ import net.minecraft.entity.passive.VillagerEntity;
 import net.minecraft.entity.vehicle.ChestMinecartEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import net.minecraft.nbt.*;
 import net.minecraft.util.Hand;
 import net.minecraft.util.math.BlockPos;
 
@@ -126,7 +129,20 @@ public class WorldGenUtils {
         desert_pyramid
     }
 
-    public static BlockPos locateFeature(Feature feature, BlockPos center) {
+    public static BlockPos locateFeature(Cubiomes.StructureType cfeature, BlockPos center) {
+				Feature feature = switch (cfeature) {
+					case Treasure -> Feature.buried_treasure;
+					case Mansion -> Feature.mansion;
+					case Stronghold -> Feature.stronghold;
+					case Fortress -> Feature.nether_fortress;
+					case Monument -> Feature.ocean_monument;
+					case Bastion -> Feature.bastion_remnant;
+					case End_City -> Feature.end_city;
+					case Village -> Feature.village;
+					case Mineshaft -> Feature.mineshaft;
+					case Desert_Pyramid -> Feature.desert_pyramid;
+					default -> null;
+				};
         Seed seed = Seeds.get().getSeed();
         BlockPos pos = null;
         if (!checkIfInDimension(getDimension(feature))) {
@@ -330,12 +346,14 @@ public class WorldGenUtils {
         return new BlockPos(pos.getX(), pos.getY(), pos.getZ());
     }
 
+    // TODO: check this lmao
     private static boolean isValidMap(Feature feature, ItemStack stack) {
-        if (!stack.hasNbt()) return false;
-        if (!stack.getNbt().contains("display")) return false;
-        NbtCompound displayTag = stack.getNbt().getCompound("display");
-        if (!displayTag.contains("Name")) return false;
-        String nameTag = displayTag.getString("Name");
+        if (stack.getComponents().isEmpty()) return false;
+        if (!stack.getComponents().contains(DataComponentTypes.MAP_DECORATIONS)) return false;
+        MapDecorationsComponent displayTag = stack.getDefaultComponents().get(DataComponentTypes.MAP_DECORATIONS);
+
+        if (!displayTag.toString().contains("Name")) return false;
+        String nameTag = String.valueOf(displayTag.decorations().get("Name"));
         if (!nameTag.contains("translate")) return false;
 
         if (feature == Feature.buried_treasure) {
@@ -349,15 +367,17 @@ public class WorldGenUtils {
     }
 
     private static BlockPos getMapMarker(ItemStack stack) {
-        if (!stack.hasNbt()) return null;
-        if (!stack.getNbt().contains("Decorations")) return null;
-        NbtList decorationsTag = stack.getNbt().getList("Decorations", NbtElement.COMPOUND_TYPE);
-        if (decorationsTag.size() < 1) return null;
-        NbtCompound iconTag = decorationsTag.getCompound(0);
+        if (stack.getComponents().isEmpty()) return null;
+        if (!stack.getDefaultComponents().contains(DataComponentTypes.MAP_DECORATIONS)) return null;
+
+        MapDecorationsComponent decorationsTag = stack.get(DataComponentTypes.MAP_DECORATIONS);
+        if (decorationsTag.decorations().isEmpty()) return null;
+        MapDecorationsComponent.Decoration iconTag = decorationsTag.decorations().get(0);
+        // check this
         return new BlockPos(
-            (int)iconTag.getDouble("x"),
-            (int)iconTag.getDouble("y"),
-            (int)iconTag.getDouble("z")
+            (int)iconTag.x(),
+            0,
+            (int)iconTag.z()
         );
     }
 }
